backend/pyproject.toml
[project]
name = "monthly-budget-backend"
version = "0.1.0"
description = "Monthly Budget Backend API"
requires-python = ">=3.12"
authors = [
    {name = "Team", email = "team@example.com"}
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "gunicorn>=21.2.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "structlog>=24.1.0",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.0",
    "PyJWT>=2.8.0",
    "google-auth>=2.25.0",
    "httpx>=0.25.0",
    "requests>=2.28.0",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
